# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bunkarr API",
    "version": "1",
    "description": "Bunkarr REST API. Authenticate with the API key from Settings > General in the X-Api-Key header (or the apikey query parameter), or with a login session cookie. Errors are {\"message\": \"...\"}: 400 invalid input, 401 not authenticated, 403 refused (a cross-site browser request, an action that needs a login session, a directory Bunkarr cannot read), 404 unknown id, 409 conflict with the current state (including safety refusals that need a confirmation or a mounted share), 502 Plex failed, 503 unavailable (the backup routes of a server without backup services, or a request stopped as the server shuts down), 500 anything else. Secrets (Plex tokens, Apprise URLs) are write-only. Job-starting endpoints answer 202 with the queued Job. Sizes are bytes, times RFC 3339.",
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    }
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "apiKeyHeader": []
    },
    {
      "apiKeyQuery": []
    },
    {
      "session": []
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      },
      "apiKeyQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "apikey"
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bunkarr_session"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "JobType": {
        "type": "string",
        "enum": [
          "scan",
          "sync",
          "plexdb_backup",
          "retention",
          "verify"
        ]
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "completed",
          "completed_with_warnings",
          "failed",
          "cancelled"
        ]
      },
      "JobTrigger": {
        "type": "string",
        "enum": [
          "schedule",
          "manual",
          "webhook",
          "resume",
          "startup"
        ]
      },
      "JobParams": {
        "type": "object",
        "description": "What a job works on; zero values are left out.",
        "properties": {
          "destinationId": {
            "type": "integer"
          },
          "sourceIds": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "integrationId": {
            "type": "integer"
          },
          "allowChanges": {
            "type": "boolean",
            "description": "Run the changes the mass-change guard would hold (S10b)."
          }
        }
      },
      "Progress": {
        "type": "object",
        "required": [
          "filesTotal",
          "filesDone",
          "bytesTotal",
          "bytesDone",
          "bytesPerSec",
          "etaSeconds"
        ],
        "properties": {
          "phase": {
            "type": "string",
            "description": "scanning, planning, copying, verifying, backing-up, finishing"
          },
          "filesTotal": {
            "type": "integer"
          },
          "filesDone": {
            "type": "integer"
          },
          "bytesTotal": {
            "type": "integer"
          },
          "bytesDone": {
            "type": "integer"
          },
          "currentFile": {
            "type": "string"
          },
          "bytesPerSec": {
            "type": "number",
            "description": "EWMA over 10 s"
          },
          "etaSeconds": {
            "type": "integer"
          }
        }
      },
      "Job": {
        "type": "object",
        "required": [
          "id",
          "type",
          "status",
          "trigger",
          "dryRun",
          "params",
          "attempt",
          "progress",
          "stats",
          "warnings",
          "summary",
          "queuedAt",
          "startedAt",
          "finishedAt"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "type": {
            "$ref": "#/components/schemas/JobType"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "trigger": {
            "$ref": "#/components/schemas/JobTrigger"
          },
          "dryRun": {
            "type": "boolean"
          },
          "params": {
            "$ref": "#/components/schemas/JobParams"
          },
          "attempt": {
            "type": "integer",
            "minimum": 1
          },
          "progress": {
            "$ref": "#/components/schemas/Progress"
          },
          "stats": {
            "type": [
              "object",
              "null"
            ],
            "description": "Runner-specific counters, e.g. sync: filesPlanned, filesCopied, filesUpdated, filesMoved, filesAdopted, filesLinked, filesRetained, filesDisplaced, filesHeld, filesFailed, filesSkipped, bytesPlanned, bytesCopied, durationMs."
          },
          "warnings": {
            "type": "integer"
          },
          "summary": {
            "type": "string"
          },
          "error": {
            "type": "string",
            "description": "Why the job failed (secrets redacted)."
          },
          "queuedAt": {
            "type": "string",
            "format": "date-time"
          },
          "startedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "finishedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
...
      "Schedule": {
        "type": "object",
        "required": [
          "id",
          "jobType",
          "params",
          "description",
          "cron",
          "enabled",
          "lastRunAt",
          "nextRunAt",
          "blockedReason"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "jobType": {
            "$ref": "#/components/schemas/JobType"
          },
          "params": {
            "$ref": "#/components/schemas/JobParams"
          },
          "description": {
            "type": "string"
          },
          "cron": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "lastRunAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "nextRunAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "null when disabled or blocked"
          },
          "blockedReason": {
            "type": "string",
            "description": "Why the scheduler refuses this schedule's jobs, e.g. 'destination \"UNAS\" is disabled' (its destination or Plex server is disabled; Run now answers 409 too); empty when they can run"
          }
        }
      },
      "ScheduleUpdate": {
        "type": "object",
        "properties": {
          "cron": {
            "type": "string",
            "description": "Empty keeps the current expression."
          },
          "enabled": {
            "type": "boolean",
            "description": "Missing keeps the current flag."
          }
        }
      },
...
      "ScheduleRunRequest": {
        "type": "object",
        "properties": {
          "dryRun": {
            "type": "boolean",
            "description": "Queue a dry run (a preview): nothing is written or deleted (S9)"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid input",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid credentials",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "No such object",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Internal error (message redacted of secrets)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "ServiceUnavailable": {
        "description": "No backup services are configured on this server, or the request was stopped (the server is shutting down)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "parameters": {
      "id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "minimum": 1
        }
      },
...
    }
  },
  "paths": {
...
    "/schedules": {
      "get": {
        "operationId": "listSchedules",
        "summary": "List schedules with their next run",
        "responses": {
          "200": {
            "description": "Schedules",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Schedule"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": [
          "Jobs"
        ]
      }
    },
    "/schedules/{id}": {
      "put": {
        "operationId": "updateSchedule",
        "summary": "Change a schedule's cron expression or enabled flag",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Schedule"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": [
          "Jobs"
        ]
      }
    },
    "/schedules/{id}/run": {
      "post": {
        "operationId": "runSchedule",
        "summary": "Run a schedule's job now",
        "description": "Queues the schedule's job with trigger manual, also when the schedule is disabled. With dryRun it queues a dry run instead (the Preview of System → Tasks): nothing is written or deleted, the job's items show what a run would do (a retention preview queues one dry-run retention job per enabled destination, whose expire items are the retained files a run would delete), and it is not recorded as a run of the schedule (lastRunAt).",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleRunRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "The queued job (an identical queued job is returned instead of a new one)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "The schedule is blocked (blockedReason): its destination or Plex server is disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": [
          "Jobs"
        ]
      }
    },
...
  }
}